Remove catalog_convert and move funcs to data_load
